Use argparse for parsing command line
In preparation for adding a configuration file and command-line
arguments, add code to use argparse to parse the command line. This is
at this point nonfunctional with the exception that the code now
enforces exactly the correct number of arguments on the command line
rather than silently ignoring extra arguments.