Don't use storeId
The code was using the storeId field in songs to import them, when it
existed. I don't know why it was doing that, but it wasn't working for
me -- songs added via storeId simply weren't appearing in my
playlists. Removing that logic and always using the id, never the
storeId, fixed this problem. This change may not be "correct", but all
I know is that the code wasn't working for me without the change and
is working with it.